Start moving views from Creator into shared files
