+++ InvisibleMap2/State Machine/Controllers/MapNavigator.swift
@@ -98,6 +98,5 @@ class MapNavigator {
     func resetMap() {
         self.stopPathPlanning()
         self.map = nil
-        
     }
 }
