+++ InvisibleMap2/Views/NavigateMapView.swift
@@ -1,0 +1,21 @@
+//
+//  NavigateMapView.swift
+//  InvisibleMap2
+//
+//  Created by tad on 11/12/21.
+//  Copyright © 2021 Occam Lab. All rights reserved.
+//
+
+import SwiftUI
+
+struct NavigateMapView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct NavigateMapView_Previews: PreviewProvider {
+    static var previews: some View {
+        NavigateMapView()
+    }
+}
