+++ Views/ContentView.swift
@@ -9,58 +9,6 @@ import SwiftUI
 import Firebase
 import FirebaseAuth
 
-class MapDatabase: ObservableObject {
-    @Published var maps: [String] = []
-    @Published var images: [UIImage] = []
-    @Published var files: [String] = []
-    var mapsRef: DatabaseReference!
-    var storageRef: StorageReference!
-    
-    init() {
-        var userMapsPath = "maps/"
-        if Auth.auth().currentUser != nil {
-            userMapsPath = userMapsPath + String(Auth.auth().currentUser!.uid)
-        }
-        mapsRef = Database.database(url: "https://invisible-map-sandbox.firebaseio.com/").reference(withPath: userMapsPath)
-        storageRef = Storage.storage().reference()
-        
-        // Tracks any addition, change, or removal to the map database
-        self.mapsRef.observe(.childAdded) { (snapshot) -> Void in
-            self.processMap(key: snapshot.key, values: snapshot.value as! [String: Any])
-        }
-        self.mapsRef.observe(.childChanged) { (snapshot) -> Void in
-            self.processMap(key: snapshot.key, values: snapshot.value as! [String: Any])
-        }
-        self.mapsRef.observe(.childRemoved) { (snapshot) -> Void in
-            if let existingMapIndex = self.maps.firstIndex(of: snapshot.key) {
-                self.maps.remove(at: existingMapIndex)
-                self.images.remove(at: existingMapIndex)
-                self.files.remove(at: existingMapIndex)
-            }
-        }
-    }
-    
-    func processMap(key: String, values: [String: Any]) {
-        // Only include in the list if it is processed
-        if let processedMapFile = values["map_file"] as? String {
-            // TODO: pick a sensible default image
-            let imageRef = storageRef.child((values["image"] as? String) ?? "olin_library.jpg")
-            imageRef.getData(maxSize: 10*1024*1024) { imageData, error in
-                if let error = error {
-                    print(error.localizedDescription)
-                    // Error occurred
-                } else {
-                    if let data = imageData {
-                        self.images.append(UIImage(data: data)!)
-                        self.files.append(processedMapFile)
-                        self.maps.append(key)
-                    }
-                }
-            }
-        }
-    }
-}
-
 class AuthListener: ObservableObject {
     init() {
         FirebaseApp.configure()
@@ -72,7 +20,7 @@ class AuthListener: ObservableObject {
 
 struct ContentView: View {
     @ObservedObject var authListener = AuthListener()
-    @ObservedObject var mapDatabase = MapDatabase()
+    @ObservedObject var mapDatabase = FirebaseManager.createMapDatabase()
     
     var body: some View {
         if Auth.auth().currentUser == nil {
@@ -101,17 +49,19 @@ struct ContentView: View {
                         }
                     }
                     Divider()
-                    // New map button
-                    NavigationLink(
-                        destination: RecordMapView()
-                    ) {
-                        Text("New Map")
-                            .frame(width: 200, height: 40)
-                            .foregroundColor(.blue)
-                            .background(
-                                RoundedRectangle(cornerRadius: 15)
-                                    .stroke(Color.blue, lineWidth: 1))
-                    }
+                    #if IS_MAP_CREATOR
+                        // New map button
+                        NavigationLink(
+                            destination: RecordMapView()
+                        ) {
+                            Text("New Map")
+                                .frame(width: 200, height: 40)
+                                .foregroundColor(.blue)
+                                .background(
+                                    RoundedRectangle(cornerRadius: 15)
+                                        .stroke(Color.blue, lineWidth: 1))
+                        }
+                    #endif
                 }
                 .listStyle(PlainListStyle())
                 .navigationTitle("All Maps")
